+++ test/Bootstrap.php
@@ -7,6 +7,8 @@
  * @license   http://github.com/sandrokeil/interop-config/blob/master/LICENSE.md New BSD License
  */
 
+declare(strict_types = 1);
+
 // set error reporting
 error_reporting(E_ALL | E_STRICT);