+++ test/TestAsset/ConnectionMandatoryContainerIdConfiguration.php
@@ -7,6 +7,8 @@
  * @license   http://github.com/sandrokeil/interop-config/blob/master/LICENSE.md New BSD License
  */
 
+declare(strict_types = 1);
+
 namespace InteropTest\Config\TestAsset;
 
 use Interop\Config\ConfigurationTrait;
@@ -20,7 +22,7 @@ class ConnectionMandatoryContainerIdConfiguration implements RequiresConfigId, R
     /**
      * @interitdoc
      */
-    public function dimensions()
+    public function dimensions(): iterable
     {
         return ['doctrine', 'connection'];
     }
@@ -28,7 +30,7 @@ class ConnectionMandatoryContainerIdConfiguration implements RequiresConfigId, R
     /**
      * @interitdoc
      */
-    public function mandatoryOptions()
+    public function mandatoryOptions(): iterable
     {
         return ['driverClass', 'params'];
     }