+++ test/TestAsset/PackageDefaultAndMandatoryOptionsConfiguration.php
@@ -7,6 +7,8 @@
  * @license   http://github.com/sandrokeil/interop-config/blob/master/LICENSE.md New BSD License
  */
 
+declare(strict_types = 1);
+
 namespace InteropTest\Config\TestAsset;
 
 use Interop\Config\ConfigurationTrait;
@@ -21,7 +23,7 @@ class PackageDefaultAndMandatoryOptionsConfiguration implements RequiresConfig, 
     /**
      * @interitdoc
      */
-    public function dimensions()
+    public function dimensions(): iterable
     {
         return ['vendor', 'package'];
     }
@@ -29,21 +31,21 @@ class PackageDefaultAndMandatoryOptionsConfiguration implements RequiresConfig, 
     /**
      * @interitdoc
      */
-    public function defaultOptions()
+    public function defaultOptions(): array
     {
         return [
             'minLength' => 2,
-            'maxLength' => 10
+            'maxLength' => 10,
         ];
     }
 
     /**
      * @interitdoc
      */
-    public function mandatoryOptions()
+    public function mandatoryOptions(): iterable
     {
         return [
-            'callback'
+            'callback',
         ];
     }
 }